feat: input, image 컴포넌트 제작
