+++ src/components/Input.jsx
@@ -1,23 +1,17 @@
 import styled from "styled-components";
 import theme from "@styles/theme";
 
-export default function Button({ children, onClick, ...props }) {
-  return (
-    <StyledButton onClick={onClick} {...props}>
-      {children}
-    </StyledButton>
-  );
+export default function Input({ placeholder, ...props }) {
+  return <StyledInput placeholder={placeholder} {...props} />;
 }
 
-const StyledButton = styled.button`
+const StyledInput = styled.input`
   width: ${(props) => props.width || "auto"};
   height: ${(props) => props.height || "auto"};
   font-size: ${(props) => props.fontSize || "14px"};
   background-color: ${({ $background }) => $background || "transparent"};
   color: ${({ color }) => color || theme.colors.gray50};
-  border: 1px solid ${({ color }) => color || theme.colors.gray50};
-  padding: 8px 12px;
-  border-radius: 25px;
-
-  cursor: pointer;
+  border: 1px solid ${({ color }) => color || theme.colors.gray90};
+  padding: 12px 16px;
+  border-radius: 4px;
 `;
