+++ src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { ThemeProvider } from "styled-components";
 import App from "./App";
-import theme from "./styles/theme";
+import theme from "@styles/theme";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
