+++ vite.config.js
@@ -11,8 +11,10 @@ export default defineConfig({
   resolve: {
     alias: [
       { find: "@", replacement: "/src" },
+      { find: "@assets", replacement: "/src/assets" },
       { find: "@components", replacement: "/src/components" },
       { find: "@pages", replacement: "/src/pages" },
+      { find: "@styles", replacement: "/src/styles" },
     ],
   },
 });
